# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "タスみる API",
    "description": "仕様書からタスクを自動生成し、担当候補を提案するAPI",
    "version": "1.0.0"
  },
  "paths": {
    "/api/tasks/generate": {
      "post": {
        "summary": "Generate Tasks",
        "description": "仕様書とメンバー情報からタスクを自動生成・マッチングし、DBに永続化する",
        "operationId": "generate_tasks_api_tasks_generate_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_generate_tasks_api_tasks_generate_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectOutputSchema"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/tasks/{task_id}": {
      "put": {
        "summary": "Update Task",
        "description": "Kanbanのドラッグ＆ドロップやインライン編集によるタスク更新",
        "operationId": "update_task_api_tasks__task_id__put",
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Task Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TaskUpdateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/settings": {
      "get": {
        "summary": "Get Settings",
        "description": "現在の設定情報を取得する (フロントエンドの初期表示用)",
        "operationId": "get_settings_api_settings_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      },
      "post": {
        "summary": "Update Settings",
        "description": "設定画面からのLLM設定や通知設定をメモリに保持する",
        "operationId": "update_settings_api_settings_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SettingsSchema"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/jobs/{job_id}/tasks": {
      "get": {
        "tags": [
          "jobs"
        ],
        "summary": "Get Job Tasks",
        "operationId": "get_job_tasks_api_jobs__job_id__tasks_get",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskDocument"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Body_generate_tasks_api_tasks_generate_post": {
        "properties": {
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File"
          },
          "text_content": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Text Content"
          },
          "members_json": {
            "type": "string",
            "title": "Members Json",
            "description": "JSON文字列化されたメンバーリスト"
          }
        },
        "type": "object",
        "required": [
          "members_json"
        ],
        "title": "Body_generate_tasks_api_tasks_generate_post"
      },
...
      "GenerationNotes": {
        "properties": {
          "failed_item_count": {
            "type": "integer",
            "title": "Failed Item Count",
            "description": "検証に失敗し破棄された項目数",
            "default": 0
          },
          "dropped_duplicate_titles": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Dropped Duplicate Titles",
            "description": "重複として除外されたタスク名"
          }
        },
        "type": "object",
        "title": "GenerationNotes",
        "description": "生成処理で捨てられた/失敗した項目を透明化するための付帯情報"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ProjectOutputSchema": {
        "properties": {
          "project": {
            "type": "string",
            "title": "Project",
            "description": "プロジェクト名",
            "default": "名称未設定プロジェクト"
          },
          "exported_at": {
            "type": "string",
            "format": "date",
            "title": "Exported At",
            "description": "エクスポート日"
          },
          "tasks": {
            "items": {
              "$ref": "#/components/schemas/TaskSchema"
            },
            "type": "array",
            "title": "Tasks",
            "description": "生成されたタスク一覧"
          },
          "generation_notes": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/GenerationNotes"
              },
              {
                "type": "null"
              }
            ],
            "description": "直近のtasks/generate実行についての付帯情報（一覧取得時はnull）"
          }
        },
        "type": "object",
        "required": [
          "exported_at",
          "tasks"
        ],
        "title": "ProjectOutputSchema"
      },
...
      "RagSource": {
        "properties": {
          "chunk_id": {
            "type": "string",
            "title": "Chunk Id"
          },
          "document_id": {
            "type": "string",
            "title": "Document Id"
          },
          "page": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Page",
            "description": "ページ番号。プレーンテキスト仕様書にはページの概念が無いため常にNone。"
          },
          "section": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Section",
            "description": "条項見出し等。無ければNone。"
          },
          "similarity": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Similarity"
          }
        },
        "type": "object",
        "required": [
          "chunk_id",
          "document_id",
          "similarity"
        ],
        "title": "RagSource",
        "description": "RAG検索で見つかった、類似度付きの関連出典1件。\n\n既存の`SourceReference`(唯一の直接の出典)とは別物: こちらは\n「関連しそうな候補」を複数件、類似度スコア付きで保持するためのもの。"
      },
...
      "SettingsSchema": {
        "properties": {
          "provider": {
            "type": "string",
            "enum": [
              "openai",
              "anthropic",
              "ollama"
            ],
            "title": "Provider",
            "description": "LLMプロバイダ"
          },
          "api_key": {
            "type": "string",
            "title": "Api Key",
            "description": "APIキー (type=password)"
          },
          "base_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Base Url",
            "description": "Ollama選択時のみ使用するBase URL"
          },
          "slack_webhook_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Slack Webhook Url",
            "description": "Slack Webhook URL"
          },
          "teams_webhook_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Teams Webhook Url",
            "description": "Teams Webhook URL"
          }
        },
        "type": "object",
        "required": [
          "provider",
          "api_key"
        ],
        "title": "SettingsSchema"
      },
...
      "SourceReference": {
        "properties": {
          "document_id": {
            "type": "string",
            "title": "Document Id"
          },
          "page": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Page",
            "description": "ページ番号。プレーンテキスト仕様書にはページの概念が無いため常にNone。"
          },
          "section": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Section",
            "description": "条項見出し等。無ければNone。"
          },
          "paragraph": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Paragraph",
            "description": "段落/チャンクの識別子（chunk_id）。"
          },
          "source_text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Text",
            "description": "出典の原文抜粋。仕様書本文からの実在の部分文字列。"
          }
        },
        "type": "object",
        "required": [
          "document_id"
        ],
        "title": "SourceReference"
      },
      "Task": {
        "properties": {
          "id": {
            "type": "string",
            "pattern": "^TASK-\\d{3,}$",
            "title": "Id"
          },
          "requirement_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Requirement Ids"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "priority": {
            "type": "string",
            "enum": [
              "high",
              "medium",
              "low",
              "unknown"
            ],
            "title": "Priority",
            "default": "unknown"
          },
          "estimated_hours": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Estimated Hours"
          },
          "required_skills": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Required Skills"
          },
          "acceptance_criteria": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Acceptance Criteria"
          },
          "source_reference": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SourceReference"
              },
              {
                "type": "null"
              }
            ]
          },
          "confidence": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Confidence"
          },
          "related_sources": {
            "items": {
              "$ref": "#/components/schemas/RagSource"
            },
            "type": "array",
            "title": "Related Sources"
          },
          "needs_review": {
            "type": "boolean",
            "title": "Needs Review",
            "default": false
          },
          "review_reasons": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Review Reasons"
          }
        },
        "type": "object",
        "required": [
          "id",
          "title",
          "description",
          "confidence"
        ],
        "title": "Task",
        "description": "最終的な構造化タスク（tasks.json の1要素）"
      },
      "TaskDocument": {
        "properties": {
          "document_id": {
            "type": "string",
            "title": "Document Id"
          },
          "tasks": {
            "items": {
              "$ref": "#/components/schemas/Task"
            },
            "type": "array",
            "title": "Tasks"
          },
          "issues": {
            "items": {
              "$ref": "#/components/schemas/backend__pipeline__tasks__schema__ValidationIssue"
            },
            "type": "array",
            "title": "Issues"
          },
          "model": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model"
          },
          "generated_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Generated At",
            "description": "ISO8601形式の生成時刻(UTC)"
          }
        },
        "type": "object",
        "required": [
          "document_id",
          "tasks"
        ],
        "title": "TaskDocument",
        "description": "tasks.json の全体（1つの要求ドキュメントから分解されたタスク一覧）"
      },
...
      "TaskSchema": {
        "properties": {
          "task_id": {
            "type": "string",
            "title": "Task Id",
            "description": "タスクID (例: T-001)"
          },
          "title": {
            "type": "string",
            "title": "Title",
            "description": "タスク名"
          },
          "assignee": {
            "type": "string",
            "title": "Assignee",
            "description": "担当候補者名"
          },
          "skill_required": {
            "type": "string",
            "title": "Skill Required",
            "description": "必要スキル（表示用の文字列。詳細はrequired_skills参照）"
          },
          "priority": {
            "type": "string",
            "enum": [
              "high",
              "medium",
              "low",
              "unknown"
            ],
            "title": "Priority",
            "description": "優先度"
          },
          "deadline": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Deadline",
            "description": "締切日"
          },
          "source_section": {
            "type": "string",
            "title": "Source Section",
            "description": "仕様書の参照箇所（表示用。詳細はsource_chunk_id/source_excerpt参照）"
          },
          "load_pct": {
            "type": "integer",
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Load Pct",
            "description": "アサイン後の予測稼働負荷"
          },
          "status": {
            "type": "string",
            "title": "Status",
            "description": "Kanbanステータス (TODO, DOING, DONE)",
            "default": "TODO"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description",
            "description": "タスクの詳細説明"
          },
          "estimated_hours": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Estimated Hours",
            "description": "見積り工数（時間）"
          },
          "required_skills": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Required Skills",
            "description": "必要スキルのリスト"
          },
          "acceptance_criteria": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Acceptance Criteria",
            "description": "完了と判断できる条件のリスト"
          },
          "source_chunk_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Chunk Id",
            "description": "出典の仕様書チャンクID"
          },
          "source_excerpt": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Excerpt",
            "description": "出典の原文抜粋"
          },
          "needs_review": {
            "type": "boolean",
            "title": "Needs Review",
            "description": "人による確認が必要かどうか",
            "default": false
          },
          "review_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Review Reason",
            "description": "要確認の理由"
          }
        },
        "type": "object",
        "required": [
          "task_id",
          "title",
          "assignee",
          "skill_required",
          "priority",
          "source_section",
          "load_pct"
        ],
        "title": "TaskSchema"
      },
      "TaskUpdateRequest": {
        "properties": {
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "assignee": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Assignee"
          },
          "priority": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "high",
                  "medium",
                  "low",
                  "unknown"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Priority"
          },
          "source_section": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Section"
          },
          "status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Status"
          }
        },
        "type": "object",
        "title": "TaskUpdateRequest"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
      "backend__pipeline__tasks__schema__ValidationIssue": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "task_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Task Ids"
          }
        },
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "title": "ValidationIssue",
        "description": "Task Validationステージが検出した問題点"
      }
    }
  }
}